tests/prettydate: fix cast and size.
You can not put a 64 bit value in a long in a 32 bit binary.

+++ tests/libntp/prettydate.c
@@ -16,7 +16,7 @@ TEST_TEAR_DOWN(prettydate) {}
 static const uint32_t HALF = 2147483648UL;
 
 TEST(prettydate, ConstantDate) {
-    l_fp t = lfpinit(3485080800L, HALF); // 2010-06-09 14:00:00.5
+    l_fp t = lfpinit((int32_t)3485080800LL, HALF); // 2010-06-09 14:00:00.5
 
 	TEST_ASSERT_EQUAL_STRING("cfba1ce0.80000000 2010-06-09T14:00:00.500Z", gmprettydate(t));
 }
